Implement #230
