+++ src/components/parts/buttons/DownloadButton.vue
@@ -28,10 +28,16 @@
       </button>
     </div>
 
-    <div v-if="showDropdown && !shouldDirectDownload"
-      class="absolute right-0 mt-2 p-0.5 bg-surface-2 rounded shadow-lg border border-surface-3 z-50">
-      <button v-for="item in dropdownItems" :key="item.key" @click.stop="selectItem(item)"
-        class="w-full p-2 text-left text-sm rounded-xs hover:bg-surface-3 flex justify-between gap-4">
+    <div
+      v-if="showDropdown && !shouldDirectDownload"
+      class="absolute right-0 mt-2 p-0.5 bg-surface-2 rounded shadow-lg border border-surface-3 z-50"
+    >
+      <button
+        v-for="item in dropdownItems"
+        :key="item.key"
+        @click.stop="selectItem(item)"
+        class="w-full p-2 text-left text-sm rounded-xs hover:bg-surface-3 flex justify-between gap-4"
+      >
         <span class="whitespace-nowrap">{{ item.label }}</span>
         <span class="opacity-70">{{ item.extensionLabel }}</span>
       </button>
